pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "speaker-toolkit"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
    "python-pptx",
    "lxml",
    "qrcode",
    "Pillow",
    "imagehash",
    "numpy",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
